Create testFrac
A fraction class that converts 'string', int, and float into 'fraction' and do fractional addition, subtraction, division, and multiplication. It also has 'simplify' function that takes each fraction into its simplest form. This fraction class also works with <,>,<=,>=,== operator for comparison purposes. Lastly, it is able to do prefix and postfix addition and subtraction. The int main() in this file is testing for the Fraction class.